requirements.txt
prettytable
pandas 
mysql-connector-python
# sudo apt-get install python3-tk
flask
